improve test-deb-install
This PR adds support for using the DOCKER_BUILD_PKGS env var to
better help defining what packege to build. It also adds support
for the integration-daemon so we can run it as a bundle.

Signed-off-by: Paul Liljenberg <liljenberg.paul@gmail.com>

add directory test

Adds missing directory test. This helps verifying that
DOCKER_BUILD_PKGS is a directory and exists before continueing.

Signed-off-by: Paul Liljenberg <liljenberg.paul@gmail.com>

fix indent

Signed-off-by: Paul Liljenberg <liljenberg.paul@gmail.com>
Upstream-commit: 620b8250e85abbe284eb8d34799df1bb73d82d6a
Component: engine