Remove Dockerfile syntax from frozen features
As discussed on the maintainers mailinglist,
and during a maintainers review session;

We've recently made changes to the Dockerfile
syntax (e.g., `HEALTHCHECK`, `SHELL`), so
mentioning the Dockerfile syntax as a frozen
feature is no longer accurate.

Removing it from frozen features does NOT
automatically state an intent to make big
changes to the syntax; making changes
to the syntax are important decisions and
should never be taken lightly.

This change is just to indicate that we
can *accept* changes if they are meaningful,
and we're confident they can be made.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>
Upstream-commit: 07250457df03396ea61dbf3782b5f8cf1094e7e0
Component: engine